Updates Immer example in readme to prevent "Invalid left-hand side in assignment expression" (#125)
Using the code as is in this example causes Invalid left-hand side in assignment expression because we are setting the value rather than doing an increment or something. I ran across this while trying to use this example and have a codesandbox to show this breaking if needed. Please lmk if I am off base here, but this is the code that's actually working for me.